RPM Accuracy Improvement
Before this commit the RPM on the cluster itself is close to accurate at lower RPM's (e.g. 1800) however it drifts higher than the game input at higher rev ranges (for example the game input at 7000RPM will show 7600RPM on the cluster), this is universal regardless of the game being played.

This commit scales down the RPM by 0.935 which makes the cluster display the correct RPM across the rev range.